refactor: unify media controls
+ Combine audio and video control components
+ Update video play icon with SVG component
+ Misc. component renaming
