--- packages/react/src/components/Card/CardMedia/controls.js
@@ -1,65 +1,0 @@
-import styled from 'styled-components'
-
-const PLAY_BUTTON_SIZE = 22
-const PROGRESS_BAR_HEIGHT = 2
-
-const sizeScales = {
-  large: 1.75,
-  small: 0.55
-}
-
-const getSize = (base, size) => base * (sizeScales[size] || 1)
-
-export const PlayButton = styled('div')`
-  position: absolute;
-  background: #fff;
-  transform: rotate(30deg) skewX(-30deg) scale(1, 0.866);
-  top: ${({ cardSize }) =>
-    `calc(50% - ${getSize(PLAY_BUTTON_SIZE / 2, cardSize)}px)`};
-  left: ${({ cardSize }) =>
-    `calc(50% - ${getSize(PLAY_BUTTON_SIZE / 2, cardSize)}px)`};
-  z-index: 2;
-  opacity: ${({ visible }) => (visible ? 1 : 0)};
-  transition: opacity 0.15s ease-in-out;
-
-  &::before,
-  &::after {
-    content: '';
-    position: absolute;
-    background: inherit;
-    left: 0;
-    top: 0;
-  }
-
-  &,
-  &::before,
-  &::after {
-    width: ${({ cardSize }) => `${getSize(PLAY_BUTTON_SIZE, cardSize)}px`};
-    height: ${({ cardSize }) => `${getSize(PLAY_BUTTON_SIZE, cardSize)}px`};
-    border-top-right-radius: 30%;
-    backface-visibility: hidden;
-  }
-
-  &::before {
-    transform: rotate(-135deg) skewX(-45deg) scale(1.414, 0.707)
-      translate(0, -50%);
-  }
-
-  &::after {
-    transform: rotate(135deg) skewY(-45deg) scale(0.707, 1.414) translate(50%);
-  }
-`
-
-export const ProgressBar = styled('div').attrs(({ progress, visible }) => ({
-  style: {
-    width: `${progress}%` || 0,
-    opacity: visible ? 0.8 : 0
-  }
-}))`
-  position: absolute;
-  left: 0;
-  bottom: 0;
-  background: #e1e8ed;
-  height: ${({ cardSize }) => getSize(PROGRESS_BAR_HEIGHT, cardSize)}px;
-  transition: opacity 0.3s ease-in-out;
-`
